backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mysqlclient==2.2.8
pillow==12.1.1
PyJWT==2.11.0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
gunicorn==23.0.0
whitenoise==6.9.0
boto3==1.38.0
